server/pyproject.toml
[tool.poetry]
name = "featherlight"
version = "1.0"
description = "GraphQL API to create a FeatherLight Endpoint"
authors = ["Sean Aye"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
starlette = "^0.13.4"
ariadne = "^0.11.0"
httpx = "^0.13.3"
uvicorn = "^0.11.5"
aioredis = "^1.3.1"
grpcio = "^1.29.0"
grpcio-tools = "^1.29.0"
googleapis-common-protos = "^1.51.0"
argon2-cffi = "^20.1.0"
gino = "^1.0.0"
grpclib = "^0.3.1"
aiostream = "^0.4.1"
pymacaroons = "^0.13.0"
wget = "^3.2"
dulwich = "^0.19.16"

[tool.poetry.scripts]
compile = "featherlight.helpers.generate_proto:compile_protos"

[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-asyncio = "^0.12.0"
mypy = "^0.770"
black = "^19.10b0"
pylint = "^2.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
